requirements.txt
# Core PyTorch and ML dependencies
torch>=1.9.0
torchvision>=0.10.0
torch-audio>=0.9.0

# Computer vision and image processing
opencv-python>=4.5.0
Pillow>=8.0.0
scikit-image>=0.18.0

# Scientific computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Machine learning utilities
scikit-learn>=1.0.0
timm>=0.4.12
pytorch-msssim>=0.2.1

# Visualization
matplotlib>=3.3.0
seaborn>=0.11.0

# Progress bars and utilities
tqdm>=4.62.0

# Medical imaging (optional, for NIFTI support)
nibabel>=3.2.0

# For LPIPS if used
lpips>=0.1.4

# For HD-BET and brain extraction (if used)
# hd-bet  # Uncomment if needed

# For advanced training features
# torchmetrics>=0.6.0  # Uncomment if needed

# Development and debugging
ipython>=7.0.0
jupyter>=1.0.0
